+++ example/lib/examples/customizable_example/customizable_example.dart
@@ -5,8 +5,10 @@ import 'package:flutter/material.dart';
 import 'package:group_button/group_button.dart';
 
 class CustomizableExample extends StatelessWidget {
-  CustomizableExample({Key key}) : super(key: key);
-  final controller = GroupButtonController();
+  CustomizableExample({Key? key}) : super(key: key);
+  final controller = GroupButtonController(
+    onDisablePressed: (i) => debugPrint('Disable Button #$i pressed'),
+  );
   final customizableController = CustomizableExampleController();
 
   @override
@@ -18,10 +20,10 @@ class CustomizableExample extends StatelessWidget {
         builder: (context, child) {
           return Scaffold(
             appBar: AppBar(
-              shadowColor: Colors.grey[100].withOpacity(0.1),
+              shadowColor: Colors.grey[100]?.withOpacity(0.1),
               backgroundColor: Colors.white,
               title: Text(
-                'GroupButton 4.5.0',
+                'GroupButton 4.6.0',
                 style: Theme.of(context).textTheme.headline6,
               ),
             ),
@@ -33,8 +35,6 @@ class CustomizableExample extends StatelessWidget {
                     .map((i) => '${i + 1}')
                     .toList(),
                 onSelected: (i, selected) => debugPrint('Button #$i $selected'),
-                onDisablePressed: (i) =>
-                    debugPrint('Disable Button #$i pressed'),
               ),
             ),
             bottomNavigationBar: GroupButtonBottomPanel(