+++ example/lib/examples/extended_example/helper_widgets/scroll_ijector.dart
@@ -3,9 +3,9 @@ import 'package:group_button/group_button.dart';
 
 class ScrollIjector extends StatelessWidget {
   const ScrollIjector({
-    Key key,
-    this.child,
-    this.groupingType,
+    Key? key,
+    required this.child,
+    required this.groupingType,
   }) : super(key: key);
 
   final Widget child;